+++ source/gui/settingsDialogs.py
@@ -983,16 +983,18 @@ class GeneralSettingsPanel(SettingsPanel):
 			if globalVars.appArgs.secure:
 				mirrorBox.Disable()
 
-		self.preventSystemLockCombo: nvdaControls.FeatureFlagCombo = settingsSizerHelper.addLabeledControl(
-			labelText=_(
-				# Translators: This is a label for a combo-box in the general settings panel.
-				"Prevent s&ystem lock during say all or reading with braille",
-			),
-			wxCtrlClass=nvdaControls.FeatureFlagCombo,
-			keyPath=["general", "preventSystemLock"],
-			conf=config.conf,
+		self.preventDisplayTurnOffCombo: nvdaControls.FeatureFlagCombo = (
+			settingsSizerHelper.addLabeledControl(
+				labelText=_(
+					# Translators: This is a label for a combo-box in the general settings panel.
+					"Prevent &display from turning off during say all or reading with braille",
+				),
+				wxCtrlClass=nvdaControls.FeatureFlagCombo,
+				keyPath=["general", "preventDisplayTurnOff"],
+				conf=config.conf,
+			)
 		)
-		self.bindHelpEvent("preventSystemLock", self.preventSystemLockCombo)
+		self.bindHelpEvent("preventDisplayTurnOff", self.preventDisplayTurnOffCombo)
 
 	def onChangeMirrorURL(self, evt: wx.CommandEvent | wx.KeyEvent):
 		"""Show the dialog to change the update mirror URL, and refresh the dialog in response to the URL being changed."""
@@ -1115,7 +1117,7 @@ class GeneralSettingsPanel(SettingsPanel):
 			updateCheck.terminate()
 			updateCheck.initialize()
 
-		self.preventSystemLockCombo.saveCurrentValueToConf()
+		self.preventDisplayTurnOffCombo.saveCurrentValueToConf()
 
 	def onPanelActivated(self):
 		if updateCheck: